+++ packages/editor-common/web/src/pluginToolbarButtonsIds.ts
@@ -5,6 +5,7 @@ const PLUGIN_TOOLBAR_BUTTON_ID = {
   REPLACE: 'replace',
   LINK: 'link',
   SIZE: 'size',
+  SEPERATOR: 'seperator',
 };
 
 export default PLUGIN_TOOLBAR_BUTTON_ID;