+++ packages/plugin-giphy/web/src/getToolbarButtons.tsx
@@ -9,10 +9,6 @@ import { gifModals } from './constants';
 import { NodeSizeButton } from 'wix-rich-content-toolbars-ui';
 import type { PluginContainerData_Width_Type } from 'ricos-schema';
 
-const RESOLVER_IDS = {
-  SELECTED: 'IS_GIF_SELECTED',
-};
-
 export const getToolbarButtons = (config): ToolbarButton[] => {
   return [
     {
@@ -27,7 +23,13 @@ export const getToolbarButtons = (config): ToolbarButton[] => {
       ),
     },
     {
+      id: PLUGIN_TOOLBAR_BUTTON_ID.SEPERATOR,
+    },
+    {
       id: PLUGIN_TOOLBAR_BUTTON_ID.ALIGNMENT,
+    },
+    {
+      id: PLUGIN_TOOLBAR_BUTTON_ID.SEPERATOR,
     },
     {
       id: PLUGIN_TOOLBAR_BUTTON_ID.REPLACE,
@@ -53,18 +55,6 @@ export const getToolbarButtons = (config): ToolbarButton[] => {
     },
     {
       id: PLUGIN_TOOLBAR_BUTTON_ID.DELETE,
-      // type: 'toggle',
-      // config: {
-      // icon: TrashIcon,
-      // tooltip: 'Delete',
-      // command: editorCommands => editorCommands.chain().focus().deleteSelection().run(),
-      // attributes: {
-      //   visible: RESOLVER_IDS.SELECTED,
-      // },
-      // },
     },
   ];
-  // resolvers: {
-  //   [RESOLVER_IDS.SELECTED]: content => content.length === 1 && content[0].type.name === 'gif',
-  // },
 };