+++ packages/plugin-vertical-embed/web/src/getToolbarButtons.tsx
@@ -33,6 +33,9 @@ export const getToolbarButtons = (config): ToolbarButton[] => {
       },
     },
     {
+      id: PLUGIN_TOOLBAR_BUTTON_ID.SEPERATOR,
+    },
+    {
       id: PLUGIN_TOOLBAR_BUTTON_ID.DELETE,
     },
   ];