+++ packages/toolbars-ui/web/src/components/index.ts
@@ -3,3 +3,4 @@ export { ToggleButton } from './ToggleButton';
 export { ListItemSelect, ListItemSection } from './ListItem';
 export { DropdownPanel } from './DropdownPanel';
 export { AddPluginMenu, PLUGIN_MENU_MODAL_ID } from './AddPluginMenu';
+export { ToolbarButtonSeparator } from './ToolbarButtonSeparator';